選手詳細画面用の処理を分割
